+++ api/annotations.go
@@ -1,0 +1,14 @@
+package api
+
+import (
+	"errors"
+)
+
+const (
+	RoleMaster = "master"
+	RoleNode   = "node"
+
+	RoleKeyPrefix = "node-role.kubernetes.io/"
+	RoleMasterKey = RoleKeyPrefix + RoleMaster
+	RoleNodeKey = RoleKeyPrefix + RoleNode
+)
