--- api/constants.go
@@ -1,6 +1,0 @@
-package api
-
-const (
-	RoleKubernetesMaster = "kubernetes-master"
-	RoleKubernetesPool   = "kubernetes-pool"
-)
