+++ cloud/providers/aws/startupscript.go
@@ -20,7 +20,7 @@ func (cm *ClusterManager) RenderMasterStarter() (string, error) {
 		}
 	*/
 	var buf bytes.Buffer
-	err = tpl.ExecuteTemplate(&buf, api.RoleKubernetesMaster, cloud.GetTemplateData(cm.ctx, cm.cluster))
+	err = tpl.ExecuteTemplate(&buf, api.RoleMaster, cloud.GetTemplateData(cm.ctx, cm.cluster))
 	if err != nil {
 		return "", err
 	}
@@ -41,7 +41,7 @@ func (cm *ClusterManager) RenderNodeStarter() (string, error) {
 		}
 	*/
 	var buf bytes.Buffer
-	err = tpl.ExecuteTemplate(&buf, api.RoleKubernetesPool, cloud.GetTemplateData(cm.ctx, cm.cluster))
+	err = tpl.ExecuteTemplate(&buf, api.RoleNode, cloud.GetTemplateData(cm.ctx, cm.cluster))
 	if err != nil {
 		return "", err
 	}
