+++ cloud/providers/linode/im.go
@@ -50,10 +50,10 @@ func (im *instanceManager) GetInstance(md *api.InstanceStatus) (*api.Instance, e
 					return err
 				}
 				if master {
-					instance.Spec.Role = api.RoleKubernetesMaster
+					instance.Spec.Role = api.RoleMaster
 				} else {
 					instance.Name = im.cluster.Name + "-node-" + strconv.Itoa(fip.LinodeId)
-					instance.Spec.Role = api.RoleKubernetesPool
+					instance.Spec.Role = api.RoleNode
 				}
 				return nil
 			}
