+++ cloud/providers/vultr/apply.go
@@ -54,7 +54,7 @@ func (cm *ClusterManager) Apply(in *api.Cluster, dryRun bool) error {
 	// -------------------------------------------------------------------ASSETS
 	im := &instanceManager{ctx: cm.ctx, cluster: cm.cluster, conn: cm.conn, namer: cm.namer}
 
-	masterScriptId, err := im.createStartupScript(cm.cluster.Spec.MasterSKU, api.RoleKubernetesMaster)
+	masterScriptId, err := im.createStartupScript(cm.cluster.Spec.MasterSKU, api.RoleMaster)
 	if err != nil {
 		cm.cluster.Status.Reason = err.Error()
 		return errors.FromErr(err).WithContext(cm.ctx).Err()
@@ -81,7 +81,7 @@ func (cm *ClusterManager) Apply(in *api.Cluster, dryRun bool) error {
 		cm.cluster.Status.Reason = err.Error()
 		return errors.FromErr(err).WithContext(cm.ctx).Err()
 	}
-	masterInstance.Spec.Role = api.RoleKubernetesMaster
+	masterInstance.Spec.Role = api.RoleMaster
 	cm.cluster.Spec.MasterExternalIP = masterInstance.Status.PublicIP
 	cm.cluster.Spec.MasterInternalIP = masterInstance.Status.PrivateIP
 	fmt.Println("Master EXTERNAL_IP", cm.cluster.Spec.MasterExternalIP, " --- Master INTERNAL_IP", cm.cluster.Spec.MasterInternalIP)
@@ -162,7 +162,7 @@ func (cm *ClusterManager) Apply(in *api.Cluster, dryRun bool) error {
 							cm.cluster.Status.Reason = err.Error()
 							return errors.FromErr(err).WithContext(cm.ctx).Err()
 						}
-						node.Spec.Role = api.RoleKubernetesPool
+						node.Spec.Role = api.RoleNode
 						cloud.Store(cm.ctx).Instances(cm.cluster.Name).Create(node)
 					}
 				}
