+++ cloud/providers/vultr/im.go
@@ -35,9 +35,9 @@ func (im *instanceManager) GetInstance(md *api.InstanceStatus) (*api.Instance, e
 			if server.InternalIP == md.PrivateIP {
 				instance, err = im.newKubeInstance(&server)
 				if master {
-					instance.Spec.Role = api.RoleKubernetesMaster
+					instance.Spec.Role = api.RoleMaster
 				} else {
-					instance.Spec.Role = api.RoleKubernetesPool
+					instance.Spec.Role = api.RoleNode
 				}
 				return
 			}
